First, the frequency of a port should be using size_t as its values
since edges are limited to size_t anyways. Also fixes an off-by-one when
polling edges.

* remotes/origin/dev/fix-frequency-usage:
  Start peeking at the beginning of the edge
  Use size_t for frequencies

+++ src/vistk/pipeline/process.h
@@ -91,7 +91,7 @@ class VISTK_PIPELINE_EXPORT process
     /// The type for the type of data on a port.
     typedef std::string port_type_t;
     /// The type for the component of a frequency.
-    typedef uint64_t frequency_component_t;
+    typedef size_t frequency_component_t;
     /// The type for the frequency of data on a port.
     typedef boost::rational<frequency_component_t> port_frequency_t;
     /// The type for a flag on a port.
